chore: move TestContainers config to a separate file
