+++ docker.go
@@ -30,7 +30,6 @@ import (
 	"github.com/docker/docker/pkg/jsonmessage"
 	"github.com/docker/go-connections/nat"
 	"github.com/google/uuid"
-	"github.com/magiconair/properties"
 	"github.com/moby/term"
 	specs "github.com/opencontainers/image-spec/specs-go/v1"
 
@@ -746,14 +745,6 @@ func (p *DockerProvider) SetClient(c client.APIClient) {
 
 var _ ContainerProvider = (*DockerProvider)(nil)
 
-// or through Decode
-type TestContainersConfig struct {
-	Host           string `properties:"docker.host,default="`
-	TLSVerify      int    `properties:"docker.tls.verify,default=0"`
-	CertPath       string `properties:"docker.cert.path,default="`
-	RyukPrivileged bool   `properties:"ryuk.container.privileged,default=false"`
-}
-
 func NewDockerClient() (cli *client.Client, host string, tcConfig TestContainersConfig, err error) {
 	tcConfig = configureTC()
 
@@ -799,40 +790,6 @@ func NewDockerClient() (cli *client.Client, host string, tcConfig TestContainers
 	defer cli.Close()
 
 	return cli, host, tcConfig, nil
-}
-
-// configureTC reads from testcontainers properties file, if it exists
-// it is possible that certain values get overridden when set as environment variables
-func configureTC() TestContainersConfig {
-	config := TestContainersConfig{}
-
-	applyEnvironmentConfiguration := func(config TestContainersConfig) TestContainersConfig {
-		ryukPrivilegedEnv := os.Getenv("TESTCONTAINERS_RYUK_CONTAINER_PRIVILEGED")
-		if ryukPrivilegedEnv != "" {
-			config.RyukPrivileged = ryukPrivilegedEnv == "true"
-		}
-
-		return config
-	}
-
-	home, err := os.UserHomeDir()
-	if err != nil {
-		return applyEnvironmentConfiguration(config)
-	}
-
-	tcProp := filepath.Join(home, ".testcontainers.properties")
-	// init from a file
-	properties, err := properties.LoadFile(tcProp, properties.UTF8)
-	if err != nil {
-		return applyEnvironmentConfiguration(config)
-	}
-
-	if err := properties.Decode(&config); err != nil {
-		fmt.Printf("invalid testcontainers properties file, returning an empty Testcontainers configuration: %v\n", err)
-		return applyEnvironmentConfiguration(config)
-	}
-
-	return applyEnvironmentConfiguration(config)
 }
 
 // BuildImage will build and image from context and Dockerfile, then return the tag
